pyproject.toml
[project]
name = "risk-assessment-workflow"
version = "0.1.0"
description = "Agentic workflow for software risk assessment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0,<1.0.0",
    "langchain>=0.3.0,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "langchain-google-genai>=2.0.0,<2.2.0",
    "google-generativeai>=0.8.0,<0.9.0",
    "google-ai-generativelanguage>=0.6.0,<0.7.0",
    "tavily-python>=0.5.0",
    "typer>=0.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "protobuf>=3.20.0,<4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
risk-assess = "risk_assessment.main:app"

